app/requirements.txt
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
certifi==2022.6.15
charset-normalizer==2.1.1
greenlet==1.1.3
h11==0.13.0
idna==3.3
outcome==1.2.0
psycopg2==2.9.3
PySocks==1.7.1
python-dotenv==0.21.0
requests==2.28.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.41
SQLAlchemy-Utils==0.38.3
tqdm==4.64.1
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.12
webdriver-manager==3.8.3
wsproto==1.2.0
